// ecommerce/src/payload/payload-types.ts
export interface Media {
  id: string;
  alt: string;
  caption?:
    | {
        [k: string]: unknown;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}
// ...
export interface Destination {
  id: string;
  title: string;
  description: string;
  image?: string | Media | null;
  location?: string | null;
  updatedAt: string;
  createdAt: string;
}